Add VisionDataset data format
This commit adds the VisionDataset data format, that can be used
to load images into RG. The class is very similar to an InMemory
dataset, the difference being that it supports map, filter, and
update operations on columns of images.

The code has been reasonably tested, but some bugs may have
slipped. Feedback welcome!