Add vision unit tests
This commit adds unit tests for VisionDataset, and also fixes some
bugs that were found.

More specifically, the images weren't being loaded when running the
filter function, the which resulted in weird bugs. Function
_inspect_filter_function fixes this behaviour by always loading the
images in the first row before calling _inspect_function.

Function append of VisionDataset had not been implemented. This has
been fixed.

 the images are now loaded before _inspect_function
runs on a filter function. The append method of the VisionDataset
had been overlooked in the previous commits and is now implemented.