Implement `to_dataloader` for VisionDataset and replaced VisionFolder with TorchDataset. (#36)
In RG workflows that involve training or evaluation loops outside of RG, it's useful if an RG dataset
can spit out a dataloader, without asking the user to define a torch dataset. For example:
```
        dataset = Dataset(...)
        for img, target in dataset.to_dataloader(
            keys=["img_path", "label"],
            batch_size=16,
            num_workers=12
        ):
            out = model(img)
            loss = loss(out, target)
            ...
```
This functionality is now implemented for Dataset with `_dataset_fmt=="vision"`.
If we want this functionality, for other formats, it will need to be implemented. Currently,
a `NotImplementedError` is raised.
Other changes:
- Noted need to install extras for pytest in CONTRIBUTING.md
- Optional dependencies `torchvision` and `PIL` are now imported lazily, instead of being set to `None`
- `num_procs` is passed onto `map` in `filter`