Add `ActivationCachedOp` and introduce `RGImageRow` and `RGImageCol`
Some slice discovery methods operate on the intermediate activations of
a trained model (GEORGE and Blind Source Separation are examples). Since
we expect several discovery methods to rely on this functionaly, we implement
a cached operation called `ActivationCachedOp` that runs a forward pass over
each example in the dataset and stores model activations in a new column. Note
this has only be tested on `VisionDataset`.

A number of changes were made to `VisionDataset` to support `ActivationCachedOp`
and future `update` and `map` based operations over vision datasets. Most notably,
`VisionDataset.__getitem__(index)` now returns instances of `RGImageRow` and
`RGImageBatch` instead of vanilla dictionaries.To the end user, these instances
should be largely indistinguishable from a batch represented by dictionary (like
those returned by`InMemoryDataset`) with the main difference being being that image
columns are loaded lazily and collated (i.e. stacked) into one tensor only when
those columns are accessed. Why do we need this? We want batches returned
by `dataset[:batch_size]` to be compatible with functions designed for `update` or
`map`. A function designed for `update` or `map` should expect a batch with
the images already loaded, since the dataloader in `map` will have already loaded
them. Previously, if we wanted to use that same function with a one-off batch (e.g.
`fn(dataset[:batch_size])`) we would first need to call `load` on each of the
images. This is a problem because (1) this assumes the user knows how to collate
the examples into a batch in a way that is consistent with the dataloader in
the dataset and (2) it requires some hacky overrides of some inherited
methods that assume `fn(dataset[:batch_size])` works (e.g. _inspect_function).
With `RGImageRow` and `RGImageBatch` is guaranteed to `fn(dataset[:batch_size])`
results consistent with what we'd get from `map(fn)`.

Additionally, one can now specify image transforms in `VisionDataset`. These are
stored in the `RGImage` instances and applied on `load.`

Other changes:
- `img_keys` are now referred to as `img_columns` throughout vision.py
- the creation of a test `VisionDataset` has been moved to `tests.testbeds`
so it can be shared with other testing modules